PyramidWriter refactoring performed
